C FFI Ready
Running cbindgen --crate orz gives okay results now. Closes #8.
