+++ Cargo.toml
@@ -4,12 +4,14 @@ build = "src/build.rs"
 edition = "2018"
 name = "orz"
 version = "1.6.0"
+
 [dependencies]
 byteorder = "1.3.2"
 log = "0.4.8"
 simplelog = "0.7.4"
 structopt = "0.3.3"
 unchecked-index = "0.2.2"
+libc = "0.2"
 
 [profile]
 [profile.release]
